Codechange: Add base() method to StrongType to allow access to the base type without casting. (#11445)
This removes the ability to explicitly cast to the base type, but the requirement
to use .base() means the conversion is still explicit.

+++ src/map.cpp
@@ -83,7 +83,7 @@ TileIndex TileAdd(TileIndex tile, TileIndexDiff add,
 
 	if (x >= Map::SizeX() || y >= Map::SizeY()) {
 		std::string message = fmt::format("TILE_ADD({}) when adding 0x{:04X} and 0x{%04X} failed",
-			exp, (uint32_t)tile, add);
+			exp, tile, add);
 #if !defined(_MSC_VER)
 		fmt::print(stderr, "{}:{} {}\n", file, line, message);
 #else
